requirements.txt
# =============================================================================
# Fraud Detection Pipeline — Python Dependencies
# =============================================================================
# pip install -r requirements.txt

# Kafka
kafka-python==2.0.2

# ML
lightgbm>=4.0.0
scikit-learn>=1.3.0
shap>=0.43.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
#tensorflow>=2.14.0

# Graph
neo4j>=5.14.0

# Monitoring
prometheus-client>=0.19.0

# Counterfactual (async worker — opsiyonel)
dice-ml>=0.11